requirements.txt

# Core dependencies for DMAIC Measure Phase
vsdx>=0.5.15
python-docx>=0.8.11
python-pptx>=0.6.21
PyMuPDF>=1.23.0
sqlalchemy>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pytest>=7.4.0
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.17.0
zipstream-ng>=1.7.0
manifestly>=0.1.0
aiosqlite>=0.19.0
pyyaml>=6.0.1
requests>=2.31.0
pathlib>=1.0.1
